Vaccine updates.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@6003 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/DrugManufacturers.cs
@@ -77,7 +77,13 @@ namespace OpenDentBusiness{
 				Meth.GetVoid(MethodBase.GetCurrentMethod(),drugManufacturerNum);
 				return;
 			}
-			string command= "DELETE FROM drugmanufacturer WHERE DrugManufacturerNum = "+POut.Long(drugManufacturerNum);
+			//validation
+			string command;
+			command="SELECT COUNT(*) FROM VaccineDef WHERE drugManufacturerNum="+POut.Long(drugManufacturerNum);
+			if(Db.GetCount(command)!="0") {
+				throw new ApplicationException(Lans.g("FormDrugUnitEdit","Cannot delete: DrugManufacturer is in use by VaccineDef."));
+			}
+			command= "DELETE FROM drugmanufacturer WHERE DrugManufacturerNum = "+POut.Long(drugManufacturerNum);
 			Db.NonQ(command);
 		}
